gcp/requirements.txt
tensorflow==2.13.1
google-cloud-storage==2.14.0
Pillow==10.2.0
functions-framework==3.4.0
